Integrated Main_Activity with Home
