+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="..\:/Projects/Android/Ezee/app/src/main/res/drawable/profile_pic.xml" value="0.2105" />
         <entry key="..\:/Projects/Android/Ezee/app/src/main/res/drawable/star.xml" value="0.2105" />
         <entry key="..\:/Projects/Android/Ezee/app/src/main/res/layout/activity_auth.xml" value="0.27291666666666664" />
+        <entry key="..\:/Projects/Android/Ezee/app/src/main/res/layout/activity_home.xml" value="0.27291666666666664" />
         <entry key="..\:/Projects/Android/Ezee/app/src/main/res/layout/activity_main.xml" value="0.25885416666666666" />
         <entry key="..\:/Projects/Android/Ezee/app/src/main/res/layout/drawer_header.xml" value="0.27291666666666664" />
         <entry key="..\:/Projects/Android/Ezee/app/src/main/res/menu/drawer_menu.xml" value="0.25" />
