+++ app/src/main/java/com/example/ezee/MainActivity.java
@@ -1,86 +1,32 @@
 package com.example.ezee;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
-
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.Window;
-import android.view.WindowManager;
+import android.view.View;
+import android.widget.Button;
 import android.widget.Toast;
 
-import androidx.appcompat.widget.Toolbar;
+public class MainActivity extends AppCompatActivity{
 
-import com.google.android.material.navigation.NavigationView;
-
-public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
-
-    DrawerLayout drawerLayout;
-    NavigationView navigationView;
-    Toolbar toolbar;
+    Button admin;
+    Button student;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-//        Window g = getWindow();
-//        g.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
-
         setContentView(R.layout.activity_main);
 
-        drawerLayout = findViewById(R.id.drawer_layout);
-        navigationView = findViewById(R.id.nav_view);
-        toolbar = findViewById(R.id.toolbar);
-
-        setSupportActionBar(toolbar);
-
-        navigationView.bringToFront();
-        navigationView.setItemTextAppearance(R.style.Menu_text_style);
-        navigationView.setCheckedItem(R.id.nav_timetable);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawerLayout, toolbar, R.string.nav_drawer_open, R.string.nav_drawer_close);
-        drawerLayout.addDrawerListener(toggle);
-        toggle.syncState();
-
-        navigationView.setNavigationItemSelectedListener(this);
+        admin = findViewById(R.id.admin);
+        student = findViewById(R.id.student);
     }
 
-    @Override
-    public void onBackPressed() {
-        if (drawerLayout.isDrawerOpen(GravityCompat.START)) {
-            drawerLayout.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
+    public void admin(View view){
+        Toast.makeText(this, "Logged in as Admin", Toast.LENGTH_SHORT).show();
     }
 
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem menu_item) {
-        switch (menu_item.getItemId()) {
-            case R.id.nav_timetable:
-                break;
-            case R.id.nav_practice:
-                Toast.makeText(this, "Practice", Toast.LENGTH_SHORT).show();
-                Intent auth = new Intent(getApplicationContext(), auth.class);
-                startActivity(auth);
-                break;
-            case R.id.nav_syllabus:
-                Toast.makeText(this, "Syllabus", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.nav_notification:
-                Toast.makeText(this, "Notifications", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.nav_feedback:
-                Toast.makeText(this, "Feedback", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.nav_logout:
-                Toast.makeText(this, "Logoout", Toast.LENGTH_SHORT).show();
-                break;
-        }
-        drawerLayout.closeDrawer(GravityCompat.START);
-        return true;
+    public void student(View view){
+        Intent home = new Intent(this, HomeActivity.class);
+        startActivity(home);
     }
 }
