+++ app/src/main/res/values/styles.xml
@@ -10,14 +10,14 @@
         <item name="colorSecondary">@color/teal_200</item>
         <item name="colorSecondaryVariant">@color/teal_700</item>
         <item name="colorOnSecondary">@color/black</item>
-        <!-- Set statusbar color -->
+        <!-- Set status color -->
         <item name="android:statusBarColor">@color/purple_700</item>
     </style>
 
     <style name="AppTheme.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
-        <item name="android:statusBarColor">#00000000</item>
+        <item name="android:statusBarColor">@color/purple_700</item>
     </style>
 
     <style name="Menu_text_style" parent="android:Widget.TextView">
